Delete the files & images

+++ controllers/stuff.js
@@ -48,7 +48,19 @@ exports.modifyThing = (req, res, next) => {
 }
 
 exports.deleteThing = (req, res, next) => {
-    Thing.deleteOne({_id: req.params.id})
-        .then(() => res.status(200).json('Objet supprimé !'))
-        .catch(error => res.status(400).json({error}));
+    Thing.findOne({_id: req.params.id})
+        .then(thing => {
+            if (thing.userId !== req.auth.userId) {
+                res.status(401).json({message: 'Not authorized'});
+            } else {
+                const filename = thing.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, (err) => {
+                    Thing.deleteOne({_id: req.params.id})
+                        .then(() => res.status(200).json({message: 'Objet supprimé !'}))
+                        .catch(error => res.status(400).json({error}));
+
+                });
+            }
+        })
+        .catch(error => res.status(500).json({error}));
 }
